+++ cmd/gitaly/main.go
@@ -10,20 +10,10 @@ import (
 
 	"gitlab.com/gitlab-org/gitaly/internal/config"
 	"gitlab.com/gitlab-org/gitaly/internal/connectioncounter"
-	"gitlab.com/gitlab-org/gitaly/internal/helper/fieldextractors"
-	"gitlab.com/gitlab-org/gitaly/internal/service"
-	"gitlab.com/gitlab-org/gitaly/internal/service/middleware/panichandler"
-	"gitlab.com/gitlab-org/gitaly/internal/service/middleware/sentryhandler"
+	"gitlab.com/gitlab-org/gitaly/internal/server"
 
-	"github.com/grpc-ecosystem/go-grpc-middleware"
-	"github.com/grpc-ecosystem/go-grpc-middleware/logging/logrus"
-	"github.com/grpc-ecosystem/go-grpc-middleware/tags"
-	"github.com/grpc-ecosystem/go-grpc-prometheus"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
-
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/reflection"
 )
 
 var version string
@@ -60,7 +50,7 @@ func validateConfig() error {
 		return fmt.Errorf("Must set $GITALY_SOCKET_PATH or $GITALY_LISTEN_ADDR")
 	}
 
-	return config.ValidateStorages()
+	return config.Validate()
 }
 
 // registerServerVersionPromGauge registers a label with the current server version
@@ -74,39 +64,6 @@ func registerServerVersionPromGauge() {
 
 	prometheus.MustRegister(gitlabBuildInfoGauge)
 	gitlabBuildInfoGauge.Set(1)
-}
-
-func newGRPCServer() *grpc.Server {
-	logrusEntry := log.NewEntry(log.StandardLogger())
-	grpc_logrus.ReplaceGrpcLogger(logrusEntry)
-
-	ctxTagOpts := []grpc_ctxtags.Option{
-		grpc_ctxtags.WithFieldExtractor(fieldextractors.RepositoryFieldExtractor),
-	}
-
-	server := grpc.NewServer(
-		grpc.StreamInterceptor(grpc_middleware.ChainStreamServer(
-			grpc_ctxtags.StreamServerInterceptor(ctxTagOpts...),
-			grpc_prometheus.StreamServerInterceptor,
-			grpc_logrus.StreamServerInterceptor(logrusEntry),
-			sentryhandler.StreamLogHandler,
-			panichandler.StreamPanicHandler, // Panic handler should remain last
-		)),
-		grpc.UnaryInterceptor(grpc_middleware.ChainUnaryServer(
-			grpc_ctxtags.UnaryServerInterceptor(ctxTagOpts...),
-			grpc_prometheus.UnaryServerInterceptor,
-			grpc_logrus.UnaryServerInterceptor(logrusEntry),
-			sentryhandler.UnaryLogHandler,
-			panichandler.UnaryPanicHandler, // Panic handler should remain last
-		)),
-	)
-
-	service.RegisterAll(server)
-	reflection.Register(server)
-
-	grpc_prometheus.Register(server)
-
-	return server
 }
 
 func main() {
@@ -144,7 +101,7 @@ func main() {
 		listeners = append(listeners, connectioncounter.New("tcp", l))
 	}
 
-	server := newGRPCServer()
+	server := server.New()
 
 	serverError := make(chan error, len(listeners))
 	for _, listener := range listeners {