+++ CHANGELOG.md
@@ -6,6 +6,8 @@ v0.12.0
   https://gitlab.com/gitlab-org/gitaly/merge_requests/179
 - Implement TreeEntry RPC
   https://gitlab.com/gitlab-org/gitaly/merge_requests/187
+- Optional token authentication
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/191
 
 v0.11.2